+++ backend/constants.js
@@ -1,0 +1,11 @@
+const PORT = 4000
+const GRID_COLUMNS = 10
+const GRID_ROWS = 10
+const BIAS_PERCENT = 0.2
+
+module.exports = {
+  PORT,
+  GRID_ROWS,
+  GRID_COLUMNS,
+  BIAS_PERCENT
+}
